autocompleteService can provide docs
Summary:
Prior to this diff:
1. autocompleteService only ever produced suggestions without documentation.
2. When the user hovers over a suggestion, and VSCode sends its completionItem/resolve message, then we fetch documentation (typically the docblock and full signature), and VSCode shows it in the autocomplete hover popup.

After this diff:
1. autocomplete is able to produce suggestions with documentation
2. When we receive completionItem/resolve, if the documentation is already present, then it's a no-op.

I'm doing this because I want to support autocomplete of builtin types. And I might as well stick in documentation in the initial response from autocomplete because (1) that's a single easy place where we know which documentation to stick in, (2) it's zero-cost.

REVIEWER INSTRUCTION: start with the change to lsp.ml. The slight complication is that we our completionItem/resolve handler must parse documentation. It's quite difficult to go back from a json documentation parameter back into the original "markedStrings list" that created it. Therefore, I sidestepped parsing.

Reviewed By: Wilfred

Differential Revision: D28776266

fbshipit-source-id: 99dafb94435f8ec81157a09a02da9d12fc6585fe