Fix bug in LowStringPtr compression
Summary: The full-width encoding for a 64-bit string pointer doesn't get tested in our environment, so I checked in a bug here and missed it. In the old code, "lo" is a uint32_t but kStringDataFlag is a uint64_t, so the | operator returns a uint64_t result. Do the | before the cast so we don't make this mistake.

Reviewed By: fredemmott

Differential Revision: D29714656

fbshipit-source-id: 7984ee7ede6ab115bfca54080850ebd28968ad53