serverTypeCheck no longer uses ServerDebug
Summary:
We have a mechanism "hh_client debug" which launches hh_server but with a special side-channel FD. Over that FD, serverTypeCheck sends a few bits of internal information about how it's doing.

We use this mechanism solely for one test, "test_method_signature_change".

I think it's not a good mechanism. It's a lot of code for just one use. I think we're better off not using this mechanism at all.

(There exist two different mechanism at the moment. We can do "hh --json" and get back detailed telemetry with internal information about how the typecheck went, but with less granularity than is needed for this test. Another option is to run hh and scrape the hh server log, but that's ugly).

(In future, hh_fanout is being rebuilt with a much better testability story, and already has lots more tests.)

So I'm just going to delete this legacy for now.

Reviewed By: dabek

Differential Revision: D29683237

fbshipit-source-id: c636eca4f9d46f7428d64edf8c6c381251058ccc