dead code - ClientDebug, DerverDebug
Summary:
clientDebug and serverDebug are both dead code.

They were only ever used by one test (removed in the previous diff), which invoked "hh_client debug" to start the server with a side-channel. clientDebug was the hh_client side of this, and serverDebug was the server side.

Note the following:
```
type 'a command =
  | Rpc of cmd_metadata * 'a t
  | Debug_DO_NOT_USE
      (** this unused constructor is part of the binary protocol
      between client and server; removing it would alter the protocol. *)
```
To see what would happen if we did remove the now-unused constructor, see following diff D29716704.

Reviewed By: dabek

Differential Revision: D29705509

fbshipit-source-id: 2f9268c54421377793ceb610d826940ac153f1ab