remove bitwise not for string in hackc
Summary:
"bitwise not" for string is a legacy operator, now typechecker doesn't accept it. Therefore hackc doesn't need to support it.

constant folding of "bitwise not" for string in hackc creates an unsafe string, it means a string may contain invalid utf8 and it will panic in `print_adata`. The reason we didn't any panic is because typechecker already rejects it.

Reviewed By: kavoor

Differential Revision: D29533045

fbshipit-source-id: de2ea8c323ef2be00f8ea7b83734b31c15eb7100