Write -1 m_data for constants that do not have a value and copy constant initializers during concretization
Summary:
# Background
paulbiss pointed out to me that KindOfUninit is an insufficient metric to judge whether an abstract constant has a default value as it is also used for constants that are initialized at runtime such as
```
class S {
  const int K = 4;
}
class C {
  const int X = S::K;
}
```
where the initializer for X is a switch statement.
```
.class {} C (7,9) {
  .const X = uninit;
  .method {}{} [private static no_injection] (7,9) 86cinit($constName) {
    .srcloc 7:1,9:33;
    CGetL $constName
    SSwitch <"X":L0 -:L1>
  L0:
    .srcloc 8:17,8:17;
    ClsCnsD "K" "S"
    .srcloc 7:1,9:33;
    RetC
  L1:
    .srcloc 7:1,9:33;
    String "Could not find initializer for "
    CGetL $constName
    String " in 86cinit"
    ConcatN 3
    Fatal Runtime
  }
}
```

In practice, this meant that

```
abstract class A {
  abstract const int X = S::K;
}
class C extends A {} // concretize X
```
would segfault HHVM when C attempts to look up a nonexistent 86cinit while a more benign example
```
class C extends A { // C has its own 86cinit
  const int Y = S::K;
}
```
would simply fatal with the error `Could not find initializer for X`.

# Diff

Since `m_data` is unused for abstract constants when the assembler generates the PreClass structs, I added an extra operation to write -1 (all 1s) into the m_data field in addition to setting the `m_kind` to `KindOfUninit` when the constant value is missing, and I updated the `isAbstractAndUninit` function to check m_data. Illustrated in HHAS
```
abstract const int X;
abstract const int Y = 3;
abstract const int Z = S::K;

.const X isAbstract; // m_kind = KindOfUninit, m_data == -1
.const Y isAbstract = """i:3;""";
.const Z isAbstract = uninit; // m_kind = KindOfUninit, m_data != -1
```

When a concretized constant has an uninit value, I copy in the 86cinit from the parent; This is roughly a reimplementation of trait flattening but only for constants and their initializers.

I also changed the logic for `hasConstant` to only use `isAbstract` instead of `isAbstractAndUninit`, as otherwise a concretized interface constant would claim that it is conflicting with the interface it came from.

However, that same function is used as the implementation of `ReflectionClass::hasConstant`. Because defaults for class constants are new to the language, this should have no effect. There will be a slight divergence between reflection on abstract type constants with defaults and abstract class constants with defaults; that just falls under the errata of needing to audit+minimize the callsites to `isAbstractAndUninit`. example:

```
abstract class A {
  abstract const int X = 3;
  abstract const type T = int;
}

$a = new ReflectionClass(A::class);
$a->hasConstant("X"); // false
$a->hasTypeConstant("T"); // true, though should be changed to false
```

Reviewed By: colavitam

Differential Revision: D29619639

fbshipit-source-id: 22f629c1dba21c001ada9064dad3df66343eba7b