suppress NAMING_COSTLY_ITER in known case
Summary:
It's costly to iterate over naming_sqlite, and we should never be doing it. The previous diff gathered telemetry about every place we do this.

There's currently only place where we know we do it -- namely, ServerLazyInit.post_saved_state_initialization. There's no point logging for this known place. (The point of telemetry is to gather information about unknown places where we log it).

Therefore, this diff suppresses the NAMING_COSTLY_ITER in the one known place where it's gathered.

Reviewed By: CatherineGasnier

Differential Revision: D29749007

fbshipit-source-id: b12c8575824d467934cecbe450ecee85fd6cbca4