NAMING_COSTLY_ITER
Summary:
It's costly to iterate over naming_sqlite, and we should never be doing it.

The previous diff helped gather telemetry about one known place where we're doing it (and which I am in the process of removing) -- namely, when we construct the dep->filename map Typing_deps.Files.ifiles.

Are there any other places where we do it? That's what this diff hopes to learn. It will gather telemetry about every place that does a costly iteration over naming_sqlite.

Reviewed By: CatherineGasnier

Differential Revision: D29748985

fbshipit-source-id: e66294e47d9a2664d2dc490af21260ff61065b48