Pull out assembly routine preconditions
Summary:
Certain string helpers are defined in assembly for performance. On x86, that's the hash table lookups and the hash function; on ARM, it's just the hash function (for now).

Today, we check the (complex) conditional for the assembly definitions in a half-dozen places. I'm going to change these conditions when I remove the m_data field for good, so let's move them to a central place and give them a better name and a comment. string-data-macros.h already defines NO_M_DATA, so it should define these flags, too.

Reviewed By: colavitam

Differential Revision: D29767508

fbshipit-source-id: 4effabb3f3db366a3f1e0f40021e7035ec164fd4