Refactoring.
Summary:
We are adding a "Fun" support in the future, which means the collection is not a set of symbols anymore, a symbol shows up in the "Extends X"/"Method X::"/"Type X" goes to the `symbols` collection (to decide class or interface.)

If a symbol shows up in the "Fun X", it goes to the `func_symbols` collection will be added in the future diffs, here we are refactoring the current implementation.

Similarly, we are supporting <Method, Fun>/<Type, Fun> in the future, current inference rule only works for <Method, Type>/<Type, Type>, therefore we are adding one more pre-condition `symbols(TypeX)`.

Broke down `DoReasoningTest` to two test suites.

Reviewed By: jamesjwu

Differential Revision: D29705485

fbshipit-source-id: 9e5d149685665a695f9d104bafd55c34b4c73f1a