Symbolize stub addresses in vasm printing
Summary:
Small quality of life improvement. Register the addresses of stubs
with the native function printer. This means that addresses of stubs
will symbolize to their names in vasm and HHIR printing. This doesn't
work all the time (because we might jump into the middle of a stub),
but it's a strict improvement.

Reviewed By: paulbiss, jano

Differential Revision: D29774657

fbshipit-source-id: 46fa78a5be6cf0880397aa3266883245178ef331