Run vasm-block-counts before register allocation
Summary:
It's desirable to run vasm-block-counts before register allocation. At
least for vasm-graph-color, we use the block counts to feed into our
heuristics. Therefore super accurate block counts can only make the
allocator make more accurate decisions.

There were a couple issues to resolve to make this happen. Right
vasm-block-counts runs at the end (after register allocation), and
does its own liveness analysis to determine which registers to
use. Since it's now running before register allocation, it can simply
use a Vreg instead and let the register allocator do the work. There's
one additional complication, however. The flags register isn't really
handled by the register allocator. It *could* be, but it adds a lot of
(largely unnecessary) complexity. vasm-block-counts doesn't know if
it's safe to clobber the flag register or not. For this purpose,
invent a new instruction, decqmlocknosf, which behaves exactly like
decqmlock, except it does not write to the flags register. This is
implemented by always saving the flags register, executing a
decqmlock, then restoring the old flags. As an optimization, if the SF
peepholes pass determines the flag is dead, it will lower
decqmlocknosf to a decqmlock. This all simplifies vasm-block-counts
considerably as it does not need the liveness analysis anymore.

The other major issue is that the register allocator can modify the
cfg. It can split critical edges (as typical). It also makes side
exiting instructions explicit (IE, jcci, bindjcc, and fallbackcc). It
needs to do this because if we exit the unit, we might need to insert
stack fixup instructions (and therefore need an exit block). It's
desirable to keep the cfg exactly the same as when vasm-block-counts
ran, to ensure maximum matching. To solve this, move the cfg
modifications out of the register allocator. Instead we'll just assert
these as pre-conditions. We'll split critical edges *before* running
vasm-block-counts and ensure it stays that way until after register
allocation. We'll simply prevent the usage of side exiting
instructions until the very end. This is accomplished by a flag to
optimizeJmps, and not calling optimizeExits until the very end. This
removes a considerable amount of logic from vasm-graph-color and
vasm-xls.

With these changes, vasm-block-counts can now be moved before register
allocation. As a nice side-effect, the opcode matching logic should be
even more effective, as we capture the opcodes *before* register
allocation, and thus changes due to spilling/rematerialization don't
matter.

Reviewed By: ottoni

Differential Revision: D29776463

fbshipit-source-id: f559d732bf310e9625e53906a9416666c72a6c36