Disable the HPHPCapturingVout lint
Summary:
Disable the HPHPCapturingVout lint. It's only really
applicable to certain constructs during code-gen, but triggers if you
use a Vout in *any* lambda, even if its perfectly fine to capture
it. Moreover, most of the time it just triggers on non-changed code,
causing a nuisance. I cannot think of a single bug ever prevented by
this lint (in fact I'm not sure if anyone has ever changed their code
due to it period), so just disable it.

Reviewed By: jano

Differential Revision: D29778156

fbshipit-source-id: 32d5053c0c3877472e733d8ff5bc8ea5207694b0