Fix prologue TransIDs
Summary: I noticed that ProfPrologues were missing in TC prints. This was due to issues with the `transId` field in the `Translator` class.  This field was serving 2 purposes: keeping the `transId` of the translation being created, which is its index in the `transdb`, and, for `OptPrologues`, keeping the `TransID` of their corresponding `ProfPrologues` (to smash the callers after emitting optimized code). This was resulting in the `OptPrologues` reusing their corresponding `ProfPrologues` entries in the `transdb`, causing them to be missing. This diff fixes the issue by using 2 separate fields in the `Translator` class for the 2 different purposes mentioned above. This diff also adds an assert in `transdb` to avoid similar bugs in the future, and cleans a couple of small things found a long the way.

Reviewed By: ricklavoie

Differential Revision: D29725753

fbshipit-source-id: 4997037bb4c63597ec7b3e3f95aa9377adbd82df