Preliminarily delete clsmeth->vec tests 1/n
Summary:
These 18 tests assume a clsmeth can be promoted to a vec. This option is flipped in prod and will be deleted in diff 5/5.

These tests all now test the same thing.

Since they will no longer be able to be treated as vecs, we only need a couple tests testing this (see tests in diff 3/5 that test comparison and casting for clsmeths: `cast.php`, `clsmeth/
primitive-cmp.php`)

[RFC] I tried to break up this change into a stack to make diff review easier. Let me know if you have any advice as to how to make diff review easier / better ways to break it up.

Reviewed By: ricklavoie

Differential Revision: D29741606

fbshipit-source-id: 42c69045d3806ed675fc7824d7ac8f0f40a8ba36