Free KeyOrders after layout selection
Summary:
After layout selection has completed, we are able to free the KeyOrders that
are not used in StructLayouts. We separate the PersistentKeyOrders used in
structs from the profiling KeyOrders and store their backing KeyOrderDatas in a
separate map. Once layout selection has completed, we place a routine to free
all of the profiling KeyOrderDatas onto the treadmill. This ensures that all
profiling has completed by the time the associated KeyOrderDatas are dropped.

Differential Revision: D29565447

fbshipit-source-id: 9abcbcfa6dd0beff6786a3079555627d8f1a142a