Remove leftover bits of debugger RPC support
Summary:
A long time ago the debugger had a custom support for debugging RPC requests.
This was implemented by intercepting all functions with a custom builtin
fb_rpc_intercept_handler(), but that builtin was removed many years ago.

Remove the remaining logic as well, since this doesn't work and the RPC
mechanism is deprecated.

Reviewed By: oulgen

Differential Revision: D29835656

fbshipit-source-id: 2e2f27008c4d757f2716a9a8202757c96c2078a1