Remove support for global handler in fb_intercept2
Summary:
This feature is broken and unused. If we are intercepting everything, by
definition we intercept also the intercept handler itself, so calling any
function results in a stack overflow. However, due to broken handling of
s_registered_flags, the first request does not intercept anything and thus the
handler does not get called. The second request intercepts all methods called
in the first request, which does not include the handler. The third request
then stack overflows on infinite handler interception.

This feature is also undesirable, as we can't promise interception will
actually work due to inlining and hhbbc folding. In the future we may require
interceptable functions to be explicitly annotated as such.

So, let's remove it.

Reviewed By: ricklavoie

Differential Revision: D29838381

fbshipit-source-id: 58d053788e9195cc037732b22b609f304a263961