Do not free still used internals of IntervalTimer
Summary:
During request shutdown an IntervalTimer may still be referenced by something
else. If we destruct its native data and then something else during the request
shutdown decrefs it (e.g. our own m_callback recursively referring to the same
IntervalTimer), we will try to destruct its native data again and crash (e.g.
by attempting to decref the same m_callback that we just freed).

Instead, just stop() it, which will clean up the thread and unregister the
IntervalTimer from the TimerPool. The destructor won't be called unless
something else decrefs it later. If not, then the memory will be reclaimed
later by the memory manager. This is fine, as the only non-trivially
destructible parts of IntervalTimer are m_thread (to std::terminate() if the
thread is still active) and m_callback (which is part of the request memory).

Reviewed By: ricklavoie

Differential Revision: D29841568

fbshipit-source-id: e0398051c392f318956ea9a694e682a2707b3c41