Reduce memory cost of interception
Summary:
If a single request uses interception, the previous logic registered a flag for
every invoked function indicating whether it could be possibly intercepted.
This registration was maintained via a map from a symbol name to a vector of
pointers to all flags stored inside Func. This could potentially have a
non-trivial memory cost.

Instead, resolve the intercepted function at the time the intercept request is
registered and flag that function directly as maybe intercepted. We don't need
to care about later loaded functions of the same name, as they can't be loaded
inside the same request.

Reviewed By: ricklavoie

Differential Revision: D29839576

fbshipit-source-id: 6e944469cc1300af74b18b437475aea17fe9966a