Fix const qualification in hh_serialize_raw / hh_store_ocaml
Summary:
In OCaml 4.10.0, String_val() evaluates to `const char *`, reflecting that
strings in OCaml are immutable. Most changes to fix this are trivial, but I'm
splitting this one out for special consideration.

Before, when serializing an OCaml string, we would grab a pointer to the string
data in the OCaml heap and pass that pointer directly into the compression
routines.

This diff allocates a buffer and copies the string data out of the OCaml heap
into that buffer. We then pass this buffer into compression routines.

Using a buffer in both the string and serialized cases means we can share the
same pointer for both branches. Alternatively, I could cast away the const
qualifier, which I believe would be safe since we don't try to modify the data
anyway.

I opted for the copy since it feels safer in the face of future code changes.
I also scanned the hack source code for SharedMem.WithCache/NoCache
applications, and couldn't find any that actually serialize a string directly.
Maybe this case can just be deleted? That's what we've done in Flow.

Differential Revision: D29850670

fbshipit-source-id: 33187ea8af80ad5c549ebb372b6f787f1ce503b9