Remove tests depending on ocaml crate
Summary:
The version of the ocaml crate vendored into fbcode does not support OCaml
4.10.0 and later. I spent some time trying to update the crate, but it has build
rules that make assumptions about ENV and paths that we can't reasonably
provide.

It seems like the only code depending on this crate are these tests, which seem
to be testing the behavior of the crate itself.

If this is the case, then I think it's simpler to just remove these tests!

Differential Revision: D29862315

fbshipit-source-id: 3e1424561d6fae97f8187ce5619af0d42706b980