Fix GVNs DecRef insertions
Summary:
GVN needs to insert IncRefs and DecRefs when re-using values. In order
to know where to insert the IncRefs and DecRefs to keep them balanced,
it does a (fairly standard) PRE-type dataflow analysis.

However, the analysis is not quite correct. There's no killing of the
partially anticipated bits anywhere. As a result, if you're in a loop,
any re-used value will be marked as partially anticipated *everywhere*
in the loop (even before it's alive or after it's dead). This doesn't
actually cause any crashes as we don't pre-maturely free
anything. However it means we'll only insert DecRefs where we exit the
loop, causing IncRefs and DecRefs to become unbalanced. It also
extends the lifetimes of the SSATmps far longer than they need to be
(this is how I noticed it).

Fix the dataflow to deal with this properly. The key idea is that the
definition of the original value should kill the associated partially
anticipated bit. While doing this, I noticed a lot of the dataflow was
totally unnecessary. We don't need to calculated availability at all
and we don't need the anticipated bits, only partially anticipated.

Reviewed By: jano

Differential Revision: D29855334

fbshipit-source-id: 1cddede4100c8f38a50edfb08a873a8734b8a815