Added Logic Program for <Method, Fun>, <Type, Fun> dependency reasoning.
Summary:
The new "Fun" node on the right hand side is very tricky to handle with "Method" edge, <Method, Fun>, as well as, "Type" edge, <Type, Fun>.

Basically, if the "Method"/"Type" are coming from a class we can simply create an object, but if they are coming from an interface, we must add a **parameter**, which leads to a chain reaction, all caller should add an **argument** as well.

The **argument** should be a class that implements the given interface in the **parameter**.

The <Method, Fun>/<SMethod, Fun> comes with a <Type, Fun> naturally, therefore, in the `<Type, Fun>` inference rule we added negation to exclude the case where <Method, Fun>/<SMethod, Fun> exists.

Reviewed By: jamesjwu

Differential Revision: D29747541

fbshipit-source-id: b54ad23e91667a7c8375af9f4ba3f2eaea511d9a