move semantics for Remove operations
Summary:
Previously, when unsetting an element in an array, we call a helper that inc-refs the original array, and then at call sites were responsible for dec-refing the result as appropriate. We are basically moving the cleanup into remove.

The new `removeMove` operator on arrays now have the following semantics:
 - Execute the unset(k) operation
 - if something was actually unset, decref on the value that was unset
 - if there was a copy, decref the array

This diff brings our `remove` operations to be in line with the semantics in D18629174 (https://github.com/facebook/hhvm/commit/7129a3732886a5657d9e0a1e7b8904871b8e1ca4).

Reviewed By: colavitam

Differential Revision: D29721601

fbshipit-source-id: e998764b72adf02ef540cd636e6bc8d36fb9cfb7