HHBC ASTs tweak for C++: Rename id types
Summary:
`cbindgen` is a simple tool that really doesn't deal well with rust "namespaces". `hhbc_id.rs` defined several types named `Type` in sub-modules all structurally equivalent to each other. Thus, `cbindgen` ultimately generated one `Type` to cover all uses of this family in the HHBC AST. There was nothing particularly wrong with this but I think it's better engineering that,  distinct types in Rust should yield distinct types in C. It's just more faithful to the intent. Understanding `cbindgen`'s shortcomings and renaming each of these types uniquely (that is, unique to the (super)crate) addresses this.

The key changes to look at are in:
- `hhbc_id.rs`

The resulting effect on the C type definitions can be seen in:
- `hhbc_id.h` & `hhbc_ast.h`

Reviewed By: shiqicao

Differential Revision: D29860042

fbshipit-source-id: 1362afb2cdde01cd6ecc54d909fff00e9349bcf5