Fix segfault on stack overflow
Summary:
It was possible to avoid stack overflow checks altogether by entering
generators. JIT implementation of ContEnter performs a basic surprise check
that exits to interpreter on surprise and stack availability below
sSurprisePageSize plus kStackCheckPadding, but the interpreter then does not
perform stack overflow check at all and only checks surprise flags.

Let's partially fix this by adding a stack overflow check to the interpreter,
similar to the one performed during FCall.

We still don't check for stack overflow due to heavy stack use by the
generator body. This is somewhat less trivial to fix efficiently and the
problem is more hypothetical, as generators live on the heap and the
sSurprisePageSize buffer for evaluation stack is generous.

Reviewed By: oulgen

Differential Revision: D29895078

fbshipit-source-id: 9786a6e168a7d244794defdc42882b2811474c3a