Remove incorrect memory effects of some instructions
Summary:
Mark a bunch of generator related instructions as not reading `$this`. These
opcodes are given ActRec inside the Generator object (which may not even have
`$this` and if it has, these opcodes do not do anything with it) and not the
ActRec of a Generator::*() method (which has Generator object as `$this`).

We do not have an alias class for generator internals yet and currently no
optimizations that could potentially benefit from that.

Reviewed By: oulgen

Differential Revision: D29897175

fbshipit-source-id: 076839374b86ce29187401ac790db1a291ecba31