Move responsibility of updating state from ContCheck to ContEnter
Summary:
Generators are in an inconsistent state between ContCheck and ContEnter. Their
state is already set to Running, but we are not yet executing inside the body.
If anything fails during this critical region (not possible yet), the generator
will never be transitioned into state capable of freeing and an assertion or
memory corruption will occur.

Solve this by moving the responsibility of setting the Running state from
ContCheck to ContEnter opcode.

Reviewed By: oulgen

Differential Revision: D29896768

fbshipit-source-id: 602951ced196fded245a538497e53df0e4e4b21a