Fix unwinding for native frames with padding after CFA
Summary:
There is no guarantee in the System V ABI that a (16-byte aligned) C++ frame
will be immediately preceeded by the prior frame. This means that when we
invoke C++ from the JIT, the prologue for the C++ function may adjust the stack
arbitrarily, adding extra padding, then pushing %rbp and copying the return
address. In fact, this is done to allow some AVX instructions to operate on the
stack.

Previously, our unwinding logic used the CFA to obtain the saved %rbp and %rip
for a C++ frame. This is not correct. While the saved %rip must be present
here, %rbp can be saved elsewhere in the frame. GCC does this for some C++
functions under march > westmere.

Two adjustments need to be made. First, sync_regstate needs to consider both
the saved %rbp and the CFA separately, as one cannot be deduced from the other.
Indirect fixup routines are updated accordingly. Secondly, syncVMRegsWork needs
to be rewritten to use the unwinder: the %rbp chain does not tell us where the
CFA is, and so indirect fixups may use the wrong offset. This diff encompasses the
first part.

Reviewed By: ricklavoie

Differential Revision: D29878492

fbshipit-source-id: e89f8e20e598093a0b7dd33848fa78ee067c5a50