Use IThriftStruct::withDefaultValues() to construct thrift objects
Summary:
HHVM pessimizes VM reentry to functions that accept default arguments and
future changes will pessimize it further. Thrift extension is the only hot
place doing it, calling constructor of IThriftStructs which accept plenty of
arguments with default values.

Usage of these constructors is also deprecated in favor of other APIs such as
fromShape() or withDefaultValues().

Let's call withDefaultValues() instead of the constructor. This avoids the
pessimization of VM reentry and as an added bonus it avoids writing the default
values on the stack, because the jitted version of withDefaultValues() inlines
the call to the constructor and just writes these default values directly to
the object fields. It also knows the types statically, so when writing a null,
only writes a single byte instead of the whole TypedValue.

Reviewed By: mkatsevVR

Differential Revision: D29900675

fbshipit-source-id: 2b48c80e8a95c39d5f9295d4fdf81a00fc0bd1f2