Remove prefix argument from Config_file getters
Summary:
The Config_file module provides an optional argument `prefix` to its getter functions. This is intended to make accessing several keys with a common prefix a little easier.

I'd like to port the representation of Config_file to Rust, so I'd like to keep it as simple as possible. To that end, I'm suggesting that we remove this prefix-adding functionality.

Reviewed By: ljw1004

Differential Revision: D29401649

fbshipit-source-id: bb967e586d0e06d4b838cc307b941037592e2285