Codemod HHVM tests broken by turning on math coercion notices
Summary: While this may technically violate the spirit of some of the tests, the tradeoff here is that when, in a later diff in the stack, we rip out the logic that allowed for coercions, we gain much more than we lose by having a number of tests that may no longer serve as much purpose as before. Thus, we determine that the tradeoff in which the tests are slightly worsened in exchange for a much quicker cleanup and path to core logic fixes is worth it.

Differential Revision: D29887026

fbshipit-source-id: 662107a4aad2d747902fbafdc68dc7ee40eceffc