Added function to emit Hack code on <Method, Fun>, <Type, Fun> edges.
Summary:
For <Interface, Fun>, we can't create an object inside the function body, therefore, we are adding a <parameter, argument> pair to the function parameter list.

The **argument** should be a class that implements the given interface in the **parameter**.

Three functions to handle `creates_in_body(?, ?).` `invokes_in_body(?, ?, ?).` `has_parameter_and_argument(?, ?, ?).` by adding related data to `class_obj_set`, `type_method_set`, and `parameter_list` respectively.

To invoke a function, currently we only has the interface as the parameter, therefore, `_print_callee` constructs the object of each class has the corresponding interface as the argument that can be passed as the parameter.

Inside the function `_print_body`, we are creating the object of each class Type, hence we satisfy the <Type, Fun> requirement by adding a parameter for <Interface, Fun> and by adding an object for <Class, Fun>.

Finally, since we got all instances for "Class"/"Interface", we can invoke the method using $Type_obj->method()

Reviewed By: jamesjwu

Differential Revision: D29799469

fbshipit-source-id: 44b8852974436651abf85a23df6debdf58dc2d57