when caching names from sqlite, only cache what we were asked for
Summary:
1. `NamingHeap.Types.get_pos id` -- if the name isn't already in shmem, then it reads from sqlite, and caches both the position into Naming_TypePos, and caches the canonical name into Naming_TypeCanon.
2. `NamingHeap.Types.get_canon_name id` -- doesn't do the same; it merely caches the canonical name into Naming_TypeCanon
3. Likewise with Fun

I wonder why get_pos is asked to cache both versions? Is it just because it's easy to compute the update to Naming_TypeCanon? That doesn't seem like a good argument.

This diff changes it so that get_pos will only cache into Naming_TypePos.

I have an ulterior motive. I want to write a "get_pos_by_dephash" function which takes a dephash. It's impossible given a dephash to convert it into a lowercase dephash. So I want to separate the two as much as possible.

Reviewed By: CatherineGasnier

Differential Revision: D29893133

fbshipit-source-id: bb0fd8fe4941e9e272956e0879553da9ea2f7c1f