Remove tast check warning for containers with covariant keys
Summary:
In typing_array_access, there is the following comment when checking the type of the index expression in `array_get`:

```
dict and keyset are covariant in the key type, so subsumption lets you upcast the key type beyond ty2 to arraykey.
e.g. consider $d: dict<string,int> and $i:int and $d[$i] should actually type check because dict<string,int> <: dict<arraykey,int>
```

The implementation there follows this comment and only requires that the indexing expression is a subtype of arraykey rather than a subtype of the index type param.

However, in `invalid_index_check` we then treat dict and keyset in the same way as all other keyed collections, requiring that the type of the indexing expression is a subtype of the relevant type param.

The upshot is that we get an error generated in the tast check for the example given in the comment. This diff changes that behaviour to make it consistent with `array_get`.

Reviewed By: CatherineGasnier

Differential Revision: D29663090

fbshipit-source-id: ddf61e388c266bf08215019600afcc2a50907923