Expected type for `Map` index expression should be subtype of key type param, not `arraykey`
Summary:
Currently, when type checking an array get on a `Map`, we perform two checks.

First, we perform a conservative check to ensure the indexing expression can be given a type which is a subtype of arraykey. This check ignores the key type parameter for the `Map`.

Second, we perform a more precise check that the indexing expression can be given a type which is a subtype of the type parameter.

If the first check gives us an error, a `Hole` is added with expected type of `arraykey`. This works nicely in cases where we have no information about the key type param:

```
function array_get_index_map_nothing(mixed $idx): void {
  $xs = Map {};
  /* HH_FIXME[4298] */
  /* HH_FIXME[4324] */
  $xs[$idx];
}
```

gives us a `Hole` on `$idx` with actual type `mixed` and expected type `arraykey`.

However, in the following case we end up with the wrong type:

```
function array_get_index_map_string(Map<string,int> $xs, mixed $idx): void {
  /* HH_FIXME[4298] */
  /* HH_FIXME[4324] */
  $xs[$idx];
}
```

i.e. we add a `Hole` with expected type of `arraykey` rather than `string`.

This diff changes the behaviour for `Map`s, only performing the more precise check based on the given key parameter type.

Related www diff is D29674328

Reviewed By: madgen

Differential Revision: D29673714

fbshipit-source-id: 3ec454fab76af62a5767139915f7bf66d79fa8f7