Add `Hole` to array get's on nullable containers not appearing as LHS of null coalesce
Summary:
Generate `Hole` on array access of nullable containers to match warnings. Uses the actual type and the most accurate non-null type available. For `option`s, this is the inner type but for e.g. `mixed`, this will be ~~`nonnull`~~ `KeyedContainer<nothing,nothing>`.

This change requires `array_get` to be extended to return (optional) errors on both the indexed and indexing expressions.

Reviewed By: madgen

Differential Revision: D29430198

fbshipit-source-id: a8f380309bdc98eb71707b6d459a171ce1fedd9b