convert `UNSAFE_CAST` `Hole`s back to `Call` during flow typing
Summary:
During flow-sensitive typing, expressions may be retyped i.e. converted back to a NAST expression. Currently, `Hole`s which have been added by an explicit call to `UNSAFE_CAST` are lost during this process.

This diff ensures such `Hole`s are rewritten to the `Call` expression that appeared in the original NAST. To support this, we add type hints to the `hole_source` `UnsafeCast` and `EnforcedCast` data constructors and use them as the type args to the corresponding function calls when rewriting to a NAST expression.

Reviewed By: madgen

Differential Revision: D29815036

fbshipit-source-id: fdd11c8272fdb7a41da748e5b7dac3e87ba6eb34