Tweak `hhas_type` types to be C++ friendly: Rename `Flags`
Summary: `cbindgen` is weak when it comes to local modules. `constraint::Flags` will get generated to just `Flags` in C++. Rename `constraint::Flags` to `constraint::ConstraintFlags`. It reduces the chances of name collision and is more descriptive anyway. Also took the chance to whack a `#[repr(C)]` on there while I was at it.

Differential Revision: D29995854

fbshipit-source-id: 5200831a28ead15edc74f9e5e7c0a9ebbf2df09a