Codemod HHVM tests broken by turning on equality coercion notices
Summary:
While this may technically violate the spirit of some of the tests, the tradeoff here is that when, in a later diff in the stack, we rip out the logic that allowed for coercions, we gain much more than we lose by having a number of tests that may no longer serve as much purpose as before. Thus, we determine that the tradeoff in which the tests are slightly worsened in exchange for a much quicker cleanup and path to core logic fixes is worth it.
This was done via running a modified version of CodemodConfigEquality over hhvm test failures

Differential Revision: D29968688

fbshipit-source-id: c4afd2fe1954eea1319ff1c546ff3f53faa6d2d6