Make most of match on class_kind stronger
Summary:
In order to improve support for enum classes and in prepration for abstract
enum classes, I'm making match on the class_kind type more resilient to
the addition of new constructors by removing most of the `| _ -> (* catch all *)`
cases.

I left a couple that would really just clutter the code.

Reviewed By: madgen

Differential Revision: D30012553

fbshipit-source-id: b74f28b6d523d0018db09ed8ddf2a4f7d9b61f00