renaming Ast_defs `class_kind` into `classish_kind`
Summary:
Renaming this type to emphasis that the `Aast.class_` structure is not only a class, but can
be an interface, a trait or an enum class. In the next diff, I'll introduce a
`Cclass` kind that will replace `Cnormal` and `Cabstract`. This is a prelude to
improve enum class support and the introduction of abstract enum class.

Reviewed By: andrewjkennedy

Differential Revision: D30039963

fbshipit-source-id: 1599640d38090045e14b42a1365ea845724c5a55