Manually fix remaining HHVM tests broken by turning on equality coercion notices
Summary: While this may technically violate the spirit of some of the tests, the tradeoff here is that when, in a later diff in the stack, we rip out the logic that allowed for coercions, we gain much more than we lose by having a number of tests that may no longer serve as much purpose as before. Thus, we determine that the tradeoff in which the tests are slightly worsened in exchange for a much quicker cleanup and path to core logic fixes is worth it.

Differential Revision: D29968685

fbshipit-source-id: 4c3fb5619ec212b98c1b7d231fbdc154ec7dd96a