JIT for StructDictUnset
Summary:
A modified version of NEUDitao's diff D29699819.

When we unset a known key in a StructDict with a known layout, we can skip the field -> slot hash table lookup. We also make a few minor optimizations here, like not emitting any code when we remove an int key from a StructDict, or a string key that we know isn't present in the layout. I doubt these cases have any perf effect.

There are other minor changes we could make to the actual hot path, like doing the "element is present" check in the JIT instead of in C++ code, but I also doubt that it matters. The field lookup is likely to be much more expensive than this check, and for most hot unset cases, the field should be present.

Reviewed By: arnabde03

Differential Revision: D30046628

fbshipit-source-id: a3495ed5958635f3eba781c5d1620dd717b3c672